Physical Activity Recognition and Remote Monitoring for Healthcare Service
Movement tracking system that recognizes remote hip exercises developed using OpenCV library and cotlin. A desktop application has been developed for users to use it in the system. A mobile application compatible with the Android operating system has been developed to display the output files created by the user. The user registers from the desktop application to the system and can enter the application by providing the necessary information to be used later. The information he entered while registering is saved to Real Time Database in Firabase. Then he selects his movement and enters the system how many times he did. Then, the user uploads the video of the user movement to the system according to the camera angle that is appropriate with the redirection in the system. Sisten only accepts videos with .Mov and .mp4 extension. The system then calculates how many times the movement has been repeated and saves the angle values ​​it measured to a file. If the same is the same as the value entered at the input, the output files are uploaded to the system. The visuals of the table showing the angle values ​​and motion openings measured to the system and the video itself are loaded. The desktop application has been developed using Python programming language. In addition, the doctor can see the previous average movement opening values ​​of the user in a table. In this way, he may have an idea of ​​the development of the patient. Doctors can also see the contact information of their patients in mobile application and receive grades of their patients.